Fix slave JVM memory calculation
This commit fixes the memory calculation for the slave JVM in the container.
Now, instead of possibly using the whole memory of the container, it
only uses the memory assigned in field "Jenkins Slave Memory in MB".

Potentially, this fixes some OOM due to the slave JVM using all the
memory.

Issue: INFRABRBA-1896