Implement ACL logic and bind Slaves to items
This commit allows Jenkins to link items to the Slave they requested which means
that only the item which requested the Slave can build on it.

Moreover, the when generating and requesting Slaves/Task the Framework to Item mapping
is now relevant.

The commit also fiexes an issue which prevented to use mutliple Frameworks within the same Mesos
master.

Issue: INFRABRBA-2415