Add USER network to networking
This commit adds the possiblity to specify a USER network.
For security reasons, right now the network name is generated using the principal and
the name of the framework (spaces are converted to dashes).

Later maybe, this "security" mechanism should be replaced by appropriate authentication/authorization
on the Mesos cluster and free to choose names for networks (has to exist however).

Issue: INFRABRBA-3615