Make sure only requests linked to item are deleted
This commits improves the behavior of requests which get removed when a
(potential) build leaves the queue. Instead of deleting a random request
which matches just the label of the, a request matching the item linked
to it gets now deleted. This ensures that the framework still provisions
requests with the same label but different linked items.

As a bonus, this commit also fixes a NPE when a pipline job completes.

Issue: INFRABRBA-3897