Fix Slave Resource Usage link isn't reliable shown in build logfiles.
This Issue was caused by the Mesos Periodic Task Reconciliation.
After a new task status was requested by the scheduler, the task status
responded by the mesos master didn't contain any information in
it's "data" field.
This field, a JSON object, is used to parse out the id of the Jenkins
Agents's docker container which is later used to generate the Slave
Resource Usage link.

Also remove "Slave resource usage is not available for this build" warning
message when no grafana dashboard link is provided in the framework's config.
Slave Resource Usage will never be available if no link is set.

Issue: INFRABRBA-3897