Introduce fitness raters
This commit introduces the concept of fitness raters, which are able
to determine the "fitness" of a offer/lease for a specific request.

Generally the fitness rating is in between 0 (unfittest) and 1
(fittest).

If it is not possible to assign a request to a lease a rater should
return -1 (not fit).